pyproject.toml
[project]
name = "ai-equity-research"
version = "1.0.0"
description = "Multi-agent AI system for equity research with convergence loop"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.18.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiohttp>=3.9",
    "yfinance>=0.2",
    "httpx>=0.26",
    "apscheduler>=3.10",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "aiosqlite>=0.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
ai-research = "scripts.run_once:main"
ai-landscape = "scripts.run_hub_daily:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "config", "scheduler", "scripts"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
